Moved checkmarkcheckbox into src\gui
